+++ runtime/stringbuf.h
@@ -144,9 +144,9 @@ rsRetVal cstrAppendCStr(cstr_t *pThis, cstr_t *pstrAppend);
 
 /* now come inline-like functions */
 #ifdef NDEBUG
-#	define cstrLen(x) ((int)((x)->iStrLen))
+#	define cstrLen(x) ((size_t)((x)->iStrLen))
 #else
-	int cstrLen(cstr_t *pThis);
+	size_t cstrLen(cstr_t *pThis);
 #endif
 #define rsCStrLen(s) cstrLen((s))
